# ASPNetServer/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "ListFun API",
    "description" : "This API will allow my frontends to communicate with my backends. Should be the same in every implementation, but I'm just testing right now anyway.",
    "version" : "0.1",
    "x-logo" : {
      "url" : ""
    }
  },
  "servers" : [ {
    "url" : "https://localhost:5000"
  } ],
  "paths" : {
...
    "/entries/add" : {
      "summary" : "Adds an entry",
      "description" : "Adds an entry to an existing Game or Movie.",
      "post" : {
        "tags" : [ ],
        "summary" : "Adds an Entry",
        "description" : "Adds an entry to an existing Game or Movie.",
        "operationId" : "addEntry",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Entry"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "200 OK."
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Entry" : {
        "required" : [ "comments", "entryid", "mediaid", "timestamp" ],
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "type" : "integer"
          },
          "comments" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "mediaid" : {
            "type" : "string"
          }
        },
        "description" : "This is the general entry. Plan is to have Games, Movies, TV Shows, whatever have an id, and every time I complete a game/movie I create an entry in this table.",
        "example" : ""
      }
    },
...
    "securitySchemes" : { }
  }
}